+++ woodwork/utils.py
@@ -8,6 +8,7 @@ import numpy as np
 import pandas as pd
 
 import woodwork as ww
+from woodwork.exceptions import TypeValidationError
 from woodwork.pandas_backport import guess_datetime_format
 
 # Dictionary mapping formats/content types to the appropriate pandas read function
@@ -214,42 +215,57 @@ def _is_url(string):
     return "http" in string
 
 
-def _reformat_to_latlong(latlong, use_list=False):
-    """Reformats LatLong columns to be tuples of floats. Uses (np.nan, np.nan) for null values."""
-    if _is_latlong_nan(latlong):
-        latlong = (np.nan, np.nan)
-
+def _reformat_to_latlong(latlong, is_koalas=False):
+    """
+    Accepts 2-tuple like values, or a single NaN like value.
+    NaN like values are replaced with np.nan.
+    """
     if isinstance(latlong, str):
+
+        nan_values = [x for x in ww.config.get_option("nan_values") if len(x)]
+        nan_values = "|".join(nan_values)
+
+        latlong = re.sub(nan_values, "None", latlong)
+
         try:
-            # Serialized latlong columns from csv or parquet will be strings, so null values will be
-            # read as the string 'nan' in pandas and Dask and 'NaN' in Koalas
-            # neither of which which is interpretable as a null value
-            if "nan" in latlong:
-                latlong = latlong.replace("nan", "None")
-            if "NaN" in latlong:
-                latlong = latlong.replace("NaN", "None")
             latlong = ast.literal_eval(latlong)
         except ValueError:
             pass
 
-    if isinstance(latlong, (tuple, list)):
+    if isinstance(latlong, (list, tuple)):
         if len(latlong) != 2:
-            raise ValueError(
+            raise TypeValidationError(
                 f"LatLong values must have exactly two values. {latlong} does not have two values."
             )
 
-        latitude, longitude = map(_to_latlong_float, latlong)
+        latitude, longitude = latlong
 
-        if use_list:
+        try:
+            latitude = _coerce_to_float(latitude)
+            longitude = _coerce_to_float(longitude)
+        except ValueError:
+            raise TypeValidationError(
+                f"LatLong values must be in decimal degrees. {latlong} does not have latitude or longitude values that can be converted to a float."
+            )
+
+        if is_koalas:
             return [latitude, longitude]
         return (latitude, longitude)
 
-    raise ValueError(
-        f"LatLongs must either be a tuple, a list, or a string representation of a tuple. {latlong} does not fit the criteria."
+    if _is_nan(latlong):
+        return np.nan
+
+    raise TypeValidationError(
+        f"""LatLong value is not properly formatted. Value must be one of the following:
+- A 2-tuple or list of 2 values representing decimal latitude or longitude values (NaN values are allowed).
+- A single NaN value.
+- A string representation of the above.
+
+{latlong} does not fit the criteria."""
     )
 
 
-def _to_latlong_float(val):
+def _coerce_to_float(val):
     """Attempts to convert a value to a float, propagating null values."""
     if _is_nan(val):
         return np.nan
@@ -258,7 +274,7 @@ def _to_latlong_float(val):
         return float(val)
     except (ValueError, TypeError):
         raise ValueError(
-            f"Latitude and Longitude values must be in decimal degrees. The latitude or longitude represented by {val} cannot be converted to a float."
+            f"The value represented by {val} cannot be converted to a float."
         )
 
 
@@ -266,28 +282,36 @@ def _is_valid_latlong_series(series):
     """Returns True if all elements in the series contain properly formatted LatLong values,
     otherwise returns False"""
     if ww.accessor_utils._is_dask_series(series):
-        series = series = series.get_partition(0).compute()
+        series = series.get_partition(0).compute()
     if ww.accessor_utils._is_koalas_series(series):
         series = series.to_pandas()
-        bracket_type = list
+        is_koalas = True
     else:
-        bracket_type = tuple
-    if series.apply(_is_valid_latlong_value, args=(bracket_type,)).all():
+        is_koalas = False
+    if series.apply(_is_valid_latlong_value, args=(is_koalas,)).all():
         return True
     return False
 
 
-def _is_valid_latlong_value(val, bracket_type=tuple):
+def _is_valid_latlong_value(val, is_koalas=False):
     """Returns True if the value provided is a properly formatted LatLong value for a
     pandas, Dask or Koalas Series, otherwise returns False."""
-    if isinstance(val, bracket_type) and len(val) == 2:
+    if isinstance(val, (list, tuple)):
+        if len(val) != 2:
+            return False
+
+        if not isinstance(val, list if is_koalas else tuple):
+            return False
+
         latitude, longitude = val
-        if isinstance(latitude, float) and isinstance(longitude, float):
-            if pd.isnull(latitude) and pd.isnull(longitude):
-                return False
-            return True
-    elif isinstance(val, float) and pd.isnull(val):
+        return isinstance(latitude, float) and isinstance(longitude, float)
+
+    if isinstance(val, float):
+        return np.isnan(val)
+
+    if is_koalas and val is None:
         return True
+
     return False